requirements.txt
anyio==4.9.0
black==25.1.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
crashtest==0.4.1
distlib==0.3.9
docutils==0.21.2
dotenv==0.9.9
dulwich==0.22.8
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
flake8==7.2.0
flit==3.12.0
flit_core==3.12.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
mccabe==0.7.0
more-itertools==10.6.0
msgpack==1.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pbs-installer==2025.4.9
pkginfo==1.12.1.2
platformdirs==4.3.7
poetry==2.1.2
poetry-core==2.1.2
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyNaCl==1.5.0
pyproject_hooks==1.2.0
python-dotenv==1.1.0
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
sniffio==1.3.1
tomli_w==1.2.0
tomlkit==0.13.2
trove-classifiers==2025.4.11.15
urllib3==2.4.0
virtualenv==20.30.0
xattr==1.1.4
zstandard==0.23.0
